Updated standard, split/add gradients, delays per object, bugfixes
The standard has been updated to include delays per event. This allows
to shift individual events within a block. In the previous standard, a
delay event in parallel to other events within a block resulted in all
other events to be shifted by the delay. This has no become obsolete.
Instead, a delay event in parallel with other events will be executed
concurrently to the other events. If the delay is longer than the
longest event in a block, this can be used to define the minimum length
of a block.

It is now possible to create "extended trapazoid" gradients. These
gradients are essentially a sequence of time/amplitude pairs. This is a
simple way to create your own shapes.

Furthermore, gradients can now be split into their "parts" using the
function "splitGradient". These gradients can then be added again using
"addGradients". These are useful tools to easily create complex, time
optimized sequences. See the new ramp-sampled EPI sequence in
writeEpiRS.m for an example of how to use this.

Gradients now do not have to be zero on the borders of blocks. However,
the following gradient must be within maxSlew of the previous amplitude.
All gradients of the final block need to end at zero.

makeTrapezoid is now also able to compute the shortest possible gradient
given all the scanner constraints.

New example sequences:
  * writeEpiRS.m: Ramp sampled EPI sequence
  * writeTrufi.m: Trufi sequence

Several bugs where fixed, introduced, and fixed in this release.